+++ server/testutils/factories/roshRisks.ts
@@ -1,0 +1,20 @@
+import type { PersonRisks } from '@approved-premises/api'
+import { RiskLevel } from '@approved-premises/ui'
+import { faker } from '@faker-js/faker/locale/en_GB'
+import { Factory } from 'fishery'
+import { DateFormats } from '../../utils/dateUtils'
+import { riskEnvelopeStatuses } from './tierEnvelopeFactory'
+
+const riskLevels: Array<RiskLevel> = ['Low', 'Medium', 'High', 'Very High']
+
+export default Factory.define<PersonRisks['roshRisks']>(() => ({
+  status: faker.helpers.arrayElement(riskEnvelopeStatuses),
+  value: {
+    overallRisk: faker.helpers.arrayElement(riskLevels),
+    riskToChildren: faker.helpers.arrayElement(riskLevels),
+    riskToPublic: faker.helpers.arrayElement(riskLevels),
+    riskToKnownAdult: faker.helpers.arrayElement(riskLevels),
+    riskToStaff: faker.helpers.arrayElement(riskLevels),
+    lastUpdated: DateFormats.dateObjToIsoDate(faker.date.past()),
+  },
+}))
